requirements.txt
django-celery==3.1.17
django-ckeditor==6.5.1
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extra-fields==3.0.2
django-utils-six==2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filetype==1.2.0
mixer==7.2.2
mysqlclient==2.2.4
pillow==10.2.0
pip-chill==1.0.3
psycopg2==2.9.9
pytest-django==4.8.0
pytest-mock==3.12.0
python-dotenv==1.0.1
requests-mock==1.11.0
stripe==8.5.0
